first implementation of module 4
Imports fixed

Finished implementation of module 4

Some minors bugs fixed

Files added to allow to see the hole folder structure

Defining the BLoC interface

EOF error fixed

Comments about EOF and imports fixed
