+++ analysis_options.yaml
@@ -8,6 +8,7 @@
 # The following line activates a set of recommended lints for Flutter apps,
 # packages, and plugins designed to encourage good coding practices.
 include: package:flutter_lints/flutter.yaml
+  package:lint/casual.yaml
 
 linter:
   # The lint rules applied to this project can be customized in the
@@ -22,6 +23,40 @@ linter:
   # `// ignore_for_file: name_of_lint` syntax on the line or in the file
   # producing the lint.
   rules:
+    - avoid_slow_async_io
+    - no_logic_in_create_state
+    - test_types_in_equals
+    - unnecessary_statements
+    - unrelated_type_equality_checks
+    - unsafe_html
+    - always_put_required_named_parameters_first
+    - always_specify_types
+    - annotate_overrides
+    - avoid_annotating_with_dynamic
+    - avoid_redundant_argument_values
+    - avoid_renaming_method_parameters
+    - avoid_unnecessary_containers
+    - await_only_futures
+    - camel_case_extensions
+    - camel_case_types
+    - cast_nullable_to_non_nullable
+    - constant_identifier_names
+    - curly_braces_in_flow_control_structures
+    - deprecated_consistency
+    - eol_at_end_of_file
+    - file_names
+    - matching_super_parameters
+    - no_leading_underscores_for_local_identifiers
+    - prefer_foreach
+    - require_trailing_commas
+    - unawaited_futures
+    - unnecessary_brace_in_string_interps
+    - unnecessary_breaks
+    - unnecessary_late
+    - unnecessary_new
+    - unnecessary_null_aware_assignments
+    - unnecessary_this
+    - use_named_constants
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
