--- lib/interface/i_my_repository.dart
@@ -1,8 +1,0 @@
-import '../shared/genre.dart';
-import '../shared/movie.dart';
-
-abstract class IMyRepository {
-  Future<List<Movie>> loadMovies();
-
-  Future<List<Genre>> loadGenres();
-}
