+++ lib/main.dart
@@ -1,36 +1,31 @@
 import 'package:flutter/material.dart';
-import 'package:movie_card/pages/home_page.dart';
-import 'pages/movie_details.dart';
-import 'shared/movie.dart';
+
+import 'src/presentation/pages/home_page.dart';
+import 'src/config/route/app_routes.dart';
+import 'src/data/models/movie_model.dart';
+import 'src/config/theme/app_themes.dart';
+import 'src/core/util/strings.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(
+    const MyApp(),
+  );
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  const MyApp({
+    super.key,
+  });
 
-  static const routeName = '/';
-
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Movie Cards',
-      initialRoute: routeName,
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
+      title: appTitle,
+      theme: AppTheme.dark,
+      routes: AppRoutes.routes,
+      home: HomePage(
+        lastMovie: MovieModel.mockMovie(),
       ),
-      routes: {
-        MovieDetails.routeName: (context) => MovieDetails(
-              movie: Movie.mockMovie(),
-            ),
-        HomePage.routeName: (context) => HomePage(
-              lastMovie: Movie.mockMovie(),
-            ),
-      },
-      home: HomePage(lastMovie: Movie.mockMovie()),
     );
   }
 }
