+++ lib/src/core/util/strings.dart
@@ -1,0 +1,3 @@
+const appTitle = 'Flutter Demo';
+
+const imageUri = 'https://image.tmdb.org/t/p/w500';
