+++ lib/src/data/models/genre_model.dart
@@ -1,0 +1,15 @@
+import '../../domain/entity/genre.dart';
+
+class GenreModel extends Genre {
+  GenreModel({
+    required super.id,
+    required super.name,
+  });
+
+  factory GenreModel.fromJson(Map<String, dynamic> json) {
+    return GenreModel(
+      id: json['id'],
+      name: json['name'],
+    );
+  }
+}
