+++ lib/src/domain/entity/cast.dart
@@ -1,0 +1,9 @@
+class Cast {
+  final String id;
+  final String profilePath;
+
+  Cast({
+    required this.id,
+    required this.profilePath,
+  });
+}
