+++ lib/src/domain/entity/genre.dart
@@ -1,0 +1,10 @@
+class Genre {
+  final int id;
+  final String name;
+
+  Genre({
+    required this.id,
+    required this.name,
+  });
+
+}
