+++ lib/src/domain/repository/i_my_repository.dart
@@ -1,0 +1,9 @@
+
+import '../entity/genre.dart';
+import '../entity/movie.dart';
+
+abstract class IMyRepository {
+  Future<List<Movie>> loadMovies();
+
+  Future<List<Genre>> loadGenres();
+}
