+++ lib/src/presentation/bloc/i_bloc.dart
@@ -1,0 +1,4 @@
+abstract class IBloc{
+  Future<void> initialize();
+  void dispose();
+}
