+++ lib/src/presentation/widgets/shared/poster.dart
